+++ src/account.js
@@ -56,6 +56,7 @@ class Account {
 
 
   get_columns() {
+    if (document.body.clientWidth <= 520) return Unit.minimal_columns
     if (document.body.clientWidth <= 800) return Unit.default_columns
     let columns = (this.data.config || {}).columns
     return (columns && columns.length) ? columns : Unit.default_columns